pseudo termino con distribucion de funcionalidades
